+++ src/libcadet/model/binding/ExtendedMobilePhaseModulatorLangmuirBinding.cpp
@@ -114,7 +114,7 @@ public:
 			throw InvalidParameterException("Not enough elements in EMPM_COMP_MODE (expected " + std::to_string(nComp) + ", got " + std::to_string(_mode.size()) + ")");
 
 		_idxModifier = -1;
-		for (int i = 0; i < _mode.size(); ++i)
+		for (int i = 0; i < static_cast<int>(_mode.size()); ++i)
 		{
 			if (_mode[i] == static_cast<int>(CompMode::Modifier))
 			{
